requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
joblib>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0
boto3>=1.28.0
python-dotenv>=1.0.0

# FastAPI backend
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# PDF generation
reportlab>=4.0.0
